requirements/prod.txt
Django==5.2
django-cleanup==9.0.0
django-tinymce==4.1.0
pillow==11.1.0
python-dotenv==1.0.1
sorl-thumbnail==12.11.0